Moved implementation common to all helper classes to a Cnab240::Helper class
