+++ lib/cnab240.rb
@@ -91,6 +91,7 @@ require "cnab240/pagamentos/v40/trailer"
 require "cnab240/arquivo/estrutura"
 require "cnab240/arquivo/builder"
 
+require "cnab240/helper/helper"
 require "cnab240/helper/pagamento_itau"
 require "cnab240/helper/pagamento_bradesco"
 require "cnab240/helper/pagamento_bb"
