+++ lib/cnab240/helper/pagamento_itau.rb
@@ -1,9 +1,5 @@
 module Cnab240
-  class PagamentoItau
-
-    include Cnab240::Filler
-
-    attr_accessor :arquivo
+  class PagamentoItau < Helper
 
     def initialize(campos = {})
       campos[:controle_banco] ||= '341'
@@ -49,12 +45,5 @@ module Cnab240
       lote << segmento_a
     end
 
-    def string
-      arquivo.string
-    end
-
-    def save_to_file(file)
-      arquivo.save_to_file(file)
-    end
   end
 end
